+++ src/main/typescript/hooks/index.ts
@@ -2,6 +2,7 @@ import registerForDragRulerReady from "./dragRuler/ready.js";
 import registerForHotbarDrop from "./hotbarDrop.js";
 import registerForInit from "./init.js";
 import registerForReady from "./ready.js";
+import registerForRenderChatLog from "./renderChatLog.js";
 import registerForRenderChatMessage from "./renderChatMessage/index.js";
 import registerForUpdateActor from "./updateActor.js";
 
@@ -11,6 +12,7 @@ export default function registerForHooks(): void {
   registerForReady();
 
   registerForHotbarDrop();
+  registerForRenderChatLog();
   registerForRenderChatMessage();
 
   registerForUpdateActor();